refactor controller package structure to handler and improve organization
