+++ internal/switauth/handler/login.go
@@ -19,24 +19,32 @@
 // THE SOFTWARE.
 //
 
-package controller
+package handler
 
 import (
 	"github.com/gin-gonic/gin"
 	"net/http"
 )
 
-func (c *AuthController) Logout(ctx *gin.Context) {
-	tokenString := ctx.GetHeader("Authorization")
-	if tokenString == "" {
-		ctx.JSON(http.StatusBadRequest, gin.H{"error": "Authorization header is missing"})
+func (c *AuthController) Login(ctx *gin.Context) {
+	var loginData struct {
+		Username string `json:"username"`
+		Password string `json:"password"`
+	}
+
+	if err := ctx.ShouldBindJSON(&loginData); err != nil {
+		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
 	}
 
-	if err := c.authService.Logout(tokenString); err != nil {
-		ctx.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+	accessToken, refreshToken, err := c.authService.Login(loginData.Username, loginData.Password)
+	if err != nil {
+		ctx.JSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
 		return
 	}
 
-	ctx.JSON(http.StatusOK, gin.H{"message": "Logged out successfully"})
+	ctx.JSON(http.StatusOK, gin.H{
+		"access_token":  accessToken,
+		"refresh_token": refreshToken,
+	})
 }
