+++ internal/switauth/handler/refresh_token.go
@@ -19,12 +19,31 @@
 // THE SOFTWARE.
 //
 
-package controller
+package handler
 
-import "github.com/gin-gonic/gin"
+import (
+	"github.com/gin-gonic/gin"
+	"net/http"
+)
 
-func HealthHandler(c *gin.Context) {
-	c.JSON(200, gin.H{
-		"message": "pong",
+func (c *AuthController) RefreshToken(ctx *gin.Context) {
+	var data struct {
+		RefreshToken string `json:"refresh_token"`
+	}
+
+	if err := ctx.ShouldBindJSON(&data); err != nil {
+		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	newAccessToken, newRefreshToken, err := c.authService.RefreshToken(data.RefreshToken)
+	if err != nil {
+		ctx.JSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
+		return
+	}
+
+	ctx.JSON(http.StatusOK, gin.H{
+		"access_token":  newAccessToken,
+		"refresh_token": newRefreshToken,
 	})
 }
