+++ internal/switauth/server/server.go
@@ -37,7 +37,7 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/innovationmech/swit/internal/switauth/client"
 	"github.com/innovationmech/swit/internal/switauth/config"
-	"github.com/innovationmech/swit/internal/switauth/controller"
+	"github.com/innovationmech/swit/internal/switauth/handler"
 	"github.com/innovationmech/swit/internal/switauth/repository"
 	"github.com/innovationmech/swit/internal/switauth/service"
 	"github.com/innovationmech/swit/internal/switauth/store"
@@ -70,7 +70,7 @@ func NewServer() (*Server, error) {
 	userClient := client.NewUserClient(sd)
 	tokenRepo := repository.NewTokenRepository(store.GetDB())
 	authService := service.NewAuthService(userClient, tokenRepo)
-	authController := controller.NewAuthController(authService)
+	authController := handler.NewAuthController(authService)
 
 	s.router = RegisterRoutes(authController)
 	return s, nil
