+++ internal/switserve/handler/stop/stop.go
