+++ internal/switserve/server/router.go
@@ -23,10 +23,10 @@ package server
 
 import (
 	"github.com/gin-gonic/gin"
-	"github.com/innovationmech/swit/internal/switserve/controller/debug"
-	"github.com/innovationmech/swit/internal/switserve/controller/health"
-	"github.com/innovationmech/swit/internal/switserve/controller/stop"
-	"github.com/innovationmech/swit/internal/switserve/controller/v1/user"
+	"github.com/innovationmech/swit/internal/switserve/handler/debug"
+	"github.com/innovationmech/swit/internal/switserve/handler/health"
+	"github.com/innovationmech/swit/internal/switserve/handler/stop"
+	"github.com/innovationmech/swit/internal/switserve/handler/v1/user"
 	"github.com/innovationmech/swit/internal/switserve/middleware"
 	"github.com/innovationmech/swit/pkg/logger"
 	"go.uber.org/zap"
